+++ infrastructure/aws/lambda/handler.py
@@ -4,7 +4,7 @@ import logging
 
 from mangum import Mangum
 
-from titiler.xarray.main import app
+from titiler.multidim.main import app
 
 logging.getLogger("mangum.lifespan").setLevel(logging.ERROR)
 logging.getLogger("mangum.http").setLevel(logging.ERROR)