+++ tests/conftest.py
@@ -1,4 +1,4 @@
-"""titiler.xarray tests configuration."""
+"""titiler.multidim tests configuration."""
 
 import pytest
 from fastapi.testclient import TestClient
@@ -7,10 +7,10 @@ from fastapi.testclient import TestClient
 @pytest.fixture
 def app(monkeypatch):
     """App fixture."""
-    monkeypatch.setenv("TITILER_XARRAY_DEBUG", "TRUE")
+    monkeypatch.setenv("TITILER_MULTIDIM_DEBUG", "TRUE")
     monkeypatch.setenv("TEST_ENVIRONMENT", "1")
 
-    from titiler.xarray.main import app
+    from titiler.multidim.main import app
 
     with TestClient(app) as client:
         yield client