+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Unreleased
+
+* Import `titiler.xarray` (from [`titiler repo`](https://github.com/developmentseed/titiler)) ([#72](https://github.com/developmentseed/titiler-xarray/pull/72))
+* Rename the package to `titiler.multidim` ([#72](https://github.com/developmentseed/titiler-xarray/pull/72)) **breaking change**
+* Drop support for kerchunk reference files ([#72](https://github.com/developmentseed/titiler-xarray/pull/72)) **breaking change**
+* Drop support for experimental `multiscale` zarr group zoom level functionality ([#72](https://github.com/developmentseed/titiler-xarray/pull/72)) **breaking change**
+* Remove default `WebMercatorQuad` tile matrix set in `/tiles`, `/tilesjson.json`, `/map` and `/WMTSCapabilities.xml` endpoints (with upgrade to `titiler.core>=0.19`) **breaking change**
+* Use `uv` for managing dependencies [#74](https://github.com/developmentseed/titiler-xarray/pull/74)
+* Slim down the Lambda asset package size [#74](https://github.com/developmentseed/titiler-xarray/pull/74)
+  * run `strip` on compiled C/C++ extensions (except `numpy.libs`)
+
 ## v0.2.0
 
 ### Improved pyramid support through group parameter
@@ -17,7 +28,7 @@
 
 ## v0.1.1
 
-Support for NetCDF and making consolidated metadata optional. See https://github.com/developmentseed/titiler-xarray/pull/39.
+Support for NetCDF and making consolidated metadata optional. See <https://github.com/developmentseed/titiler-xarray/pull/39>.
 
 [Performance results between prod (v0.1.0) and dev (unreleased)](https://github.com/developmentseed/tile-benchmarking/blob/bd1703209bbeab501f312d99fc51fda6bd419bf9/03-e2e/compare-prod-dev.ipynb).
 
@@ -26,8 +37,6 @@ Support for NetCDF and making consolidated metadata optional. See https://github
   * NetCDF Dataset: pr_day_ACCESS-CM2_historical_r1i1p1f1_gn_1950.nc
   * Unconsolidated metadata dataset: prod-giovanni-cache-GPM_3IMERGHH_06_precipitationCal
 
-
 ## v0.1.0 (2023-10-11)
 
 Initial release of the project.
-